Store file renamed to root marker
Instead of guessing the format of the include_re, store the original
file name for the module.

In theory, this could go wrong if one project checkout used one name,
module had changes that renamed it, another project checkout used the
different name, and then tried to checkin the first project. I'm not too
bothered by that scenario.