Apply include/exclude patterns to directories
Allow users to ignore examples and other nonessential content that may
be mistaken for the module.

The nata lua module contains a full-fledged demo complete with modules
using a 'init.lua'. We don't want to process those (or copy them into
our project), so we exclude them with a pattern.