checkin: Copy back all and restore deleted ignored
If ignored files exist in the project, we probably copied them manually
(can't make a perfect config -- can't account for everything). If we
modify them, we still want those changes imported.

If we're ignoring the files, then we wouldn't expect them to exist in
the target so deleting them is meaningless. Limit to deleted files so we
track changes to manually copied files.