Support copying the license file
Requires reconfiguring all project kinds.

Since most licenses require redistribution and keeping the license with
the code makes it easier to manage, make it easy to copy licenses along
with the source. Especially useful for godot where the code is in a
subfolder that's matched for the destination project, but the license is
at root.